requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jwt==1.3.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
